[refactor] AI이미지 생성 코드 리팩토링
