+++ src/main/java/com/multi/multi_semi/ai_image/controller/FrontAiImageController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 @RequiredArgsConstructor
-public class FrontImageController {
+public class FrontAiImageController {
 
     // "이미지 생성" 화면
     @GetMapping("/generate-request")
