+++ src/main/java/com/multi/multi_semi/ai_image/dto/AiImageDto.java
@@ -1,0 +1,4 @@
+package com.multi.multi_semi.ai_image.dto;
+
+
+public record AiImageDto(String memEmail, String orgUrl){}
