--- src/main/java/com/multi/multi_semi/ai_image/dto/AiImgDto.java
@@ -1,7 +1,0 @@
-package com.multi.multi_semi.ai_image.dto;
-
-
-public record AiImgDto(
-        String memEmail,
-        String orgUrl
-){}
