+++ src/main/java/com/multi/multi_semi/common/config/SecurityConfigJwt.java
@@ -90,7 +90,7 @@ public class SecurityConfigJwt {
                         .requestMatchers(HttpMethod.POST, "/api/v1/generate-status/**").hasAnyRole("ADMIN", "USER")
                         .requestMatchers("/api/v1/download-image/**").permitAll()
                         .requestMatchers("/multiimgs/**").permitAll()
-                        .requestMatchers("api/v1/weather/**").permitAll()
+                        .requestMatchers("/api/v1/weather/**").permitAll()
                         .anyRequest().authenticated())
 
                 // [ ★★★ 수정 ★★★ ]
